pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-practice"
version = "0.1.0"
description = "RAG practice project with LangChain, vector DB, and multimodal OpenAI"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.2.14",
  "langchain-openai>=0.1.7",
  "faiss-cpu>=1.8.0",
  "langchain-text-splitters>=0.2.2",
  "pypdf>=4.2.0",
  "pillow>=10.4.0",
  "python-dotenv>=1.0.1",
  "tiktoken>=0.7.0",
  "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "mypy>=1.11.0",
  "ruff>=0.6.1",
  "types-requests>=2.32.0.20240914",
]
eval = [
  "deepeval>=1.3.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unreachable = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
